WIP: Refactor into separate class/file
